Move common test functions to separate module
This resolves warnings for missing function prototypes.
